test(commit): 補助プログラムの構成変更に追従してテストを再構成
確認する内容は変えずに、
現在のソース構成へ追従させました。

`buildEditorInvocation`と`editorCommand`のテストを、
これらの実装が移った`get-editor.ts`に合わせて、
`get-editor.test.ts`へ移設しました。

`konokaEdit`が2引数化して実ファイルを読み書きするようになったため、
一時ディレクトリに`COMMIT_EDITMSG`と`patch`を用意する`setupFiles`を追加し、
`it.scoped`で実行するようにしました。

`EditorFailedError`は型付きエラーチャンネルに乗る通常のfailureなので、
`Cause.dieOption`ではなく`Cause.failureOption`で捕捉するよう修正しました。
